Implement MicronEnsemble class
Using the Micron Sonar User Manual and the Seanet DumpLog Software
Manual, the MicronSonar.py ensemble class was created to parse a
ensemble object from a list of parameters.

Additionally, skeleton code for the micron_reader file is given,
which is meant to provide the ability to parse a whole Micron
time series from a reference csv file. Also, updates to the plotting
functions used in the jupyter notebook are included.