Add function to save and open TimeSeries with CSV
Now MicronTimeSeries objects can be saved as and parsed from CSV file
format. Note that the 'MicronTimeSeries.from_csv()' function takes in
an already parsed TimeSeries object (i.e. metric units, vertical
range extracted, etc.) For raw CSV files that are logged by the
instrument directly, they must first be converted to a TimeSeries
object using the micron_reader file.

Additionally, the array size of the MicronEnsemble is not hard-coded
so that the size of the ensemble is constant throughout the mission,
regardless of the sensor configuration.