Add superclass to Ensemble and TimeSeries objects
Created a super-class MicronSonar that implements init function and
many property functions that MicronEnsemble MicronTimeSeries
inherit from. This reduces repeated code between the two classes, as
well as simplify several of the class methods as well.

Additionally, a bug was discovered with the step-size reported by
the Micron Sonar. For each resolution setting, the 'steps' variable
was a factor of two lower than the difference between two consecutive
bearing measurements. This has been fixed at parse time.

To support labeling ice data, a cropping function was added to the
TimeSeries object to select only the window of measurements that
correspond with viewing the surface (instead of the seafloor). Using
this function, labels were given to all sonar data files.