Add class-method constructors to MicronTimeSeries
Multiple class-method constructors were added to the MicronTimeSeries
class so that time series objects can be created from parsed CSV files,
from raw CSV files, from a directory of parsed CSV files, or from
incrementally adding MicronEnsembles and then converting to a DataFrame
representation. As such, the micron_reader.py file was removed because
the functionality was replaced by a new time series class method.

Also, the variable lists in the MicronSonar super class were changed
to tuples to promote immutability and also to add a 'version' attribute
to the MicronSonar class by taking the hash of 'label_list' property
(which is now a tuple).